+++ src/store/film-data/index.ts
@@ -1,0 +1,3 @@
+export * from './slice/film-data.ts';
+export * from './api-action/api-action.ts';
+export * from './selectors/selectors.ts';
