+++ src/store/genre-process/index.ts
@@ -1,0 +1,2 @@
+export * from './slice/genre-process.ts';
+export * from './selectors/selectors.ts';
