+++ src/store/genre-process/selectors/selectors.ts
@@ -1,4 +1,4 @@
 import { NameSpace } from '../../../const.ts';
-import { State } from '../../../types/state.ts';
+import { State } from '../../../types';
 
 export const getActiveGenre = (state: Pick<State, NameSpace.Genre>): string => state[NameSpace.Genre].genre;
