+++ src/store/index.ts
@@ -1,16 +1,9 @@
-import { configureStore } from '@reduxjs/toolkit';
-import { rootReducer } from './root-reducer';
-import { createAPI } from '../services/api';
-import { redirect } from './middlewares/redirect';
-
-export const api = createAPI();
-
-export const store = configureStore({
-  reducer: rootReducer,
-  middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware({
-      thunk: {
-        extraArgument: api,
-      },
-    }).concat(redirect),
-});
+export * from './middlewares/redirect.ts';
+export * from './film-data/index.ts';
+export * from './genre-process/index.ts';
+export * from './my-list-process/index.ts';
+export * from './post-review-process/index.ts';
+export * from './review-data/index.ts';
+export * from './user-process/index.ts';
+export * from './action.ts';
+export * from './root-reducer.ts';
