+++ src/store/my-list-process/selectors/selectors.test.ts
@@ -1,6 +1,6 @@
 import {describe, expect} from 'vitest';
 import {NameSpace} from '../../../const.ts';
-import {makeFakePreviewFilms} from '../../../utils/mocks.ts';
+import {makeFakePreviewFilms} from '../../../utils';
 import {getFavoriteFilmCount, getFavoriteFilms, getFavoriteFilmsDataLoading} from './selectors.ts';
 
 describe('', () => {
