+++ src/store/my-list-process/selectors/selectors.ts
@@ -1,6 +1,5 @@
 import { NameSpace } from '../../../const.ts';
-import { PreviewFilm } from '../../../types/preview-film.ts';
-import { State } from '../../../types/state.ts';
+import {PreviewFilm, State} from '../../../types';
 
 export const getFavoriteFilms = (state: Pick<State, NameSpace.MyList>): PreviewFilm[] => state[NameSpace.MyList].favoriteFilms;
 export const getFavoriteFilmsDataLoading = (state: Pick<State, NameSpace.MyList>): boolean => state[NameSpace.MyList].isFavoriteFilmsLoading;
