+++ src/store/post-review-process/api-action/api-action.ts
@@ -1,10 +1,9 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
-import { AppDispatch } from '../../../types/state.ts';
 import { AxiosInstance } from 'axios';
 import {APIRoute, AppRoute} from '../../../const.ts';
-import { ReviewAddingData } from '../../../types/review-adding-data.ts';
 import {redirectToRoute} from '../../action.ts';
 import {toast} from 'react-toastify';
+import {AppDispatch, ReviewAddingData} from '../../../types';
 
 export const postReview = createAsyncThunk<
   void,
