+++ src/store/post-review-process/selectors/selectors.ts
@@ -1,4 +1,4 @@
-import {State} from '../../../types/state.ts';
+import {State} from '../../../types';
 import {NameSpace} from '../../../const.ts';
 
 export const getFormReviewSubmitting = (state: Pick<State, NameSpace.PostingReview >): boolean => state[NameSpace.PostingReview].isFormReviewSubmitting;
