+++ src/store/root-reducer.ts
@@ -1,11 +1,11 @@
 import { combineReducers } from '@reduxjs/toolkit';
 import { NameSpace } from '../const';
-import { userProcess } from './user-process/slice/user-process.ts';
-import { filmData } from './film-data/slice/film-data.ts';
-import { genreProcess } from './genre-process/slice/genre-process.ts';
-import { reviewData } from './review-data/slice/review-data.ts';
-import { myListProcess } from './my-list-process/slice/my-list-process.ts';
-import {postingReviewProcess} from './post-review-process/slice/post-review-process.ts';
+import {filmData} from './film-data';
+import {reviewData} from './review-data';
+import {userProcess} from './user-process';
+import {myListProcess} from './my-list-process';
+import {postingReviewProcess} from './post-review-process';
+import {genreProcess} from './genre-process';
 
 export const rootReducer = combineReducers({
   [NameSpace.Genre]: genreProcess.reducer,
