+++ src/store/user-process/api-action/api-action.ts
@@ -1,14 +1,12 @@
 import {createAsyncThunk} from '@reduxjs/toolkit';
-import {AppDispatch, State} from '../../../types/state.ts';
 import {AxiosInstance} from 'axios';
-import {UserData} from '../../../types/user-data.ts';
 import {APIRoute, AppRoute} from '../../../const.ts';
-import {AuthData} from '../../../types/auth-data.ts';
 import {dropToken, saveToken} from '../../../services/token.ts';
 import {redirectToRoute} from '../../action.ts';
 import {fetchFavoriteFilmsAction} from '../../my-list-process/api-action/api-action.ts';
 import {clearMyList} from '../../my-list-process/slice/my-list-process.ts';
 import {toast} from 'react-toastify';
+import {AppDispatch, AuthData, State, UserData} from '../../../types';
 
 export const checkAuthAction = createAsyncThunk<string, undefined, {
   dispatch: AppDispatch;
