+++ src/store/user-process/slice/user-process.test.ts
@@ -1,7 +1,7 @@
 import {describe, expect} from 'vitest';
 import {AuthorizationStatus} from '../../../const.ts';
 import {userProcess} from './user-process.ts';
-import {makeFakeAvatarUrl} from '../../../utils/mocks.ts';
+import {makeFakeAvatarUrl} from '../../../utils';
 import {checkAuthAction, loginAction, logoutAction} from '../api-action/api-action.ts';
 
 describe('UserProcess slice', () => {
