+++ src/types/index.ts
@@ -1,0 +1,10 @@
+export * from './auth-data.ts';
+export * from './favorite-film-post-data.ts';
+export * from './film.ts';
+export * from './film-favorite-status.ts';
+export * from './preview-film.ts';
+export * from './promo-film.ts';
+export * from './review.ts';
+export * from './review-adding-data.ts';
+export * from './state.ts';
+export * from './user-data.ts';
