+++ src/utils/get-films-by-genre/get-films-by-genre.ts
@@ -1,9 +1,9 @@
-import { DEFAULT_GENRE } from '../../const.ts';
-import { PreviewFilm } from '../../types/preview-film.ts';
+import {PreviewFilm} from '../../types/preview-film.ts';
+import {DEFAULT_GENRE} from '../../const.ts';
 
 export const getFilmsByGenre = (films: PreviewFilm[], genre: string) => {
   if (genre === DEFAULT_GENRE) {
-    return [...films];
+    return films;
   }
   return films.filter((film) => film.genre === genre);
 };
