+++ src/utils/index.ts
@@ -1,0 +1,9 @@
+export * from './change-date-format/change-date-format.ts';
+export * from './form-regx/form-regx.ts';
+export * from './get-films-by-genre/get-films-by-genre.ts';
+export * from './get-film-rating/get-film-rating.ts';
+export * from './get-genre-list/get-genre-list.ts';
+export * from './get-run-time/get-run-time.ts';
+export * from './get-time-left/get-time-left.ts';
+export * from './mocks.ts';
+export * from './mock-components.tsx';
